update copyright year(s) if number of changed lines reaches limit
Currently years in the copyright notice are updated only if there is
some other change in it. With this commit years are also updated if
number of changed(added + deleted) lines in a per file diff reaches
a limit, which is by default 5.

Also fix handling of single year, which was updated even if it would
be the only change in the copyright notice.

Co-authored-by: Sergei Silnov <po@kumekay.com>
Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>